Create two_vectors_addition.cuda
This program defines a kernel function add that adds two arrays a and b element-wise and stores the result in an array c. It also defines an n variable that determines the size of the arrays. In main, it allocates memory on the device for the arrays, initializes host arrays a and b, copies them to the device, launches the kernel, copies the result back to the host, and prints it. Finally, it frees the memory on the device.

To compile and run this program, you need to have the CUDA toolkit installed and set up on your system. Assuming you have nvcc (the CUDA compiler) in your PATH, you can compile the program using the following command:

```nvcc -o vector_add vector_add.cu```
This will create an executable file named vector_add. You can run it using the following command in bash:

./vector_add

This should output the results of adding the two vectors. Note that the add kernel is launched with one block of n threads, so each thread handles one element of the arrays. The __global__ qualifier indicates that this function is executed on the device (GPU) and can be called from the host (CPU).